refactor(shell): split HelpModal into focused sibling files
Decompose the 687-LOC HelpModal.tsx into a slim orchestrator plus
three focused sibling modules. No behavior change.

- helpModalStyles.ts (90 LOC): CSS-in-JS STYLES object + the
  ShortcutItem/ShortcutCategory interfaces + KEY_SEPARATOR /
  formatKey / getModifierKey helpers.
- helpModalShortcutData.tsx (118 LOC): the SHORTCUT_CATEGORIES catalog
  (data only) + HelpCategoryIcon used by the data entries.
- HelpModalSections.tsx (265 LOC): all presentational sub-components
  (KeyboardKey, ShortcutRow, ShortcutCategorySection,
  MouseInteractionsSection, TouchGesturesSection, TipsSection,
  BlockedZonesSection, BinClearanceSection).
- HelpModal.tsx (269 LOC): main dialog — search/filter + tab state +
  layout shell only.
