+++ src/shell/Modals/HelpModal/HelpModal.tsx
@@ -1,187 +1,31 @@
-import { useEffect, useState, useMemo, type CSSProperties } from 'react';
-import { SHORTCUTS } from '@/core/constants';
+/**
+ * Help modal — keyboard shortcuts catalog + tips reference.
+ *
+ * Splits across:
+ *   - `helpModalStyles`         — shared style objects + key formatting helpers
+ *   - `helpModalShortcutData`   — SHORTCUT_CATEGORIES catalog (data only)
+ *   - `HelpModalSections`       — presentational sub-components
+ */
+
+import { useEffect, useState, useMemo } from 'react';
 import { useTranslation } from '@/i18n';
 import { ICON_PATHS } from '@/shared/constants/iconPaths';
-
-// Style constants to avoid recreating objects on each render
-const STYLES = {
-  // Overlay and modal
-  overlay: { backgroundColor: 'var(--overlay-dark)' } as CSSProperties,
-  modal: {
-    backgroundColor: 'var(--bg-secondary)',
-    border: '1px solid var(--border-default)',
-    borderRadius: 'var(--radius-xl)',
-    boxShadow: 'var(--shadow-xl)',
-  } as CSSProperties,
-  // Typography
-  title: {
-    fontSize: 'var(--text-2xl)',
-    fontWeight: 'var(--font-bold)',
-    color: 'var(--text-primary)',
-  } as CSSProperties,
-  sectionHeader: {
-    fontSize: 'var(--text-sm)',
-    fontWeight: 'var(--font-semibold)',
-    color: 'var(--text-tertiary)',
-    textTransform: 'uppercase',
-    letterSpacing: '0.05em',
-  } as CSSProperties,
-  textPrimary: { color: 'var(--text-primary)' } as CSSProperties,
-  textSecondary: { color: 'var(--text-secondary)' } as CSSProperties,
-  colorPrimary: { color: 'var(--color-primary)' } as CSSProperties,
-  // Containers
-  sectionContent: {
-    backgroundColor: 'var(--bg-elevated)',
-    border: '1px solid var(--border-subtle)',
-  } as CSSProperties,
-  tipsList: {
-    backgroundColor: 'var(--bg-elevated)',
-    border: '1px solid var(--border-subtle)',
-    fontSize: 'var(--text-sm)',
-    color: 'var(--text-secondary)',
-  } as CSSProperties,
-  blockedZonesContent: {
-    backgroundColor: 'var(--bg-elevated)',
-    border: '1px solid var(--border-subtle)',
-    fontSize: 'var(--text-sm)',
-    color: 'var(--text-secondary)',
-  } as CSSProperties,
-  // Button
-  buttonCompact: { minWidth: 'auto', minHeight: 'auto' } as CSSProperties,
-  // Rows
-  rowDescription: { fontSize: 'var(--text-sm)', color: 'var(--text-secondary)' } as CSSProperties,
-  rowAction: { fontSize: 'var(--text-xs)', color: 'var(--text-tertiary)' } as CSSProperties,
-} as const;
+import { STYLES, getModifierKey } from './helpModalStyles';
+import { SHORTCUT_CATEGORIES } from './helpModalShortcutData';
+import {
+  ShortcutCategorySection,
+  MouseInteractionsSection,
+  TouchGesturesSection,
+  TipsSection,
+  BlockedZonesSection,
+  BinClearanceSection,
+} from './HelpModalSections';
 
 interface HelpModalProps {
   isOpen: boolean;
   onClose: () => void;
   isTablet?: boolean;
 }
-
-// Shortcut categories with their shortcuts
-interface ShortcutItem {
-  keys: string | readonly string[];
-  descriptionKey: string; // Translation key
-  modifier?: boolean; // Whether to show Ctrl/⌘ prefix
-  shift?: boolean; // Whether to show Shift prefix
-}
-
-interface ShortcutCategory {
-  id: string;
-  nameKey: string; // Translation key
-  icon: React.ReactNode;
-  shortcuts: ShortcutItem[];
-}
-
-const KEY_SEPARATOR = '+';
-
-const getModifierKey = () => {
-  if (typeof navigator === 'undefined') return 'Ctrl';
-  const isMac = /mac/i.test(navigator.userAgent);
-  return isMac ? '⌘' : 'Ctrl';
-};
-
-const formatKey = (key: string | readonly string[]): string => {
-  if (Array.isArray(key)) {
-    return key.join(' / ');
-  }
-  return key as string;
-};
-
-// Define shortcut categories using translation keys
-const SHORTCUT_CATEGORIES: ShortcutCategory[] = [
-  {
-    id: 'general',
-    nameKey: 'help.category.general',
-    icon: <HelpCategoryIcon paths={ICON_PATHS.menu} />,
-    shortcuts: [
-      { keys: 'K', descriptionKey: 'help.shortcut.commandPalette', modifier: true },
-      { keys: formatKey(SHORTCUTS.UNDO), descriptionKey: 'common.undo', modifier: true },
-      { keys: formatKey(SHORTCUTS.REDO), descriptionKey: 'common.redo', modifier: true },
-      { keys: formatKey(SHORTCUTS.HELP), descriptionKey: 'help.shortcut.showHelp' },
-      { keys: formatKey(SHORTCUTS.ESCAPE), descriptionKey: 'help.shortcut.cancelDeselect' },
-      { keys: SHORTCUTS.TOOL_SWITCH, descriptionKey: 'help.shortcut.toolSwitch', shift: true },
-    ],
-  },
-  {
-    id: 'editing',
-    nameKey: 'help.category.editing',
-    icon: <HelpCategoryIcon paths={ICON_PATHS.edit} />,
-    shortcuts: [
-      { keys: 'D', descriptionKey: 'help.shortcut.duplicate', modifier: true },
-      { keys: formatKey(SHORTCUTS.DELETE), descriptionKey: 'help.shortcut.delete' },
-      { keys: formatKey(SHORTCUTS.ROTATE).toUpperCase(), descriptionKey: 'help.shortcut.rotate' },
-      {
-        keys: formatKey(SHORTCUTS.QUICK_LABEL).toUpperCase(),
-        descriptionKey: 'help.shortcut.quickLabel',
-      },
-      { keys: 'A', descriptionKey: 'help.shortcut.selectAll', modifier: true },
-      { keys: 'Arrow keys', descriptionKey: 'help.shortcut.nudge' },
-    ],
-  },
-  {
-    id: 'navigation',
-    nameKey: 'help.category.navigation',
-    icon: <HelpCategoryIcon paths={ICON_PATHS.navigation} />,
-    shortcuts: [
-      {
-        keys: formatKey(SHORTCUTS.LAYER_UP).toUpperCase(),
-        descriptionKey: 'help.shortcut.layerUp',
-      },
-      {
-        keys: formatKey(SHORTCUTS.LAYER_DOWN).toUpperCase(),
-        descriptionKey: 'help.shortcut.layerDown',
-      },
-      {
-        keys: formatKey(SHORTCUTS.SELECT_PREV_BIN).toUpperCase(),
-        descriptionKey: 'help.shortcut.prevBin',
-      },
-      {
-        keys: formatKey(SHORTCUTS.SELECT_NEXT_BIN).toUpperCase(),
-        descriptionKey: 'help.shortcut.nextBin',
-      },
-      {
-        keys: `${formatKey(SHORTCUTS.CATEGORY_PREV)} / ${formatKey(SHORTCUTS.CATEGORY_NEXT)}`,
-        descriptionKey: 'help.shortcut.cycleCategory',
-      },
-    ],
-  },
-  {
-    id: 'view',
-    nameKey: 'help.category.view',
-    icon: <HelpCategoryIcon paths={ICON_PATHS.eye} />,
-    shortcuts: [
-      { keys: formatKey(SHORTCUTS.ZOOM_IN), descriptionKey: 'help.shortcut.zoomIn' },
-      { keys: formatKey(SHORTCUTS.ZOOM_OUT), descriptionKey: 'help.shortcut.zoomOut' },
-      { keys: 'O', descriptionKey: 'help.shortcut.openLayoutManager', modifier: true },
-    ],
-  },
-  {
-    id: '3d-preview',
-    nameKey: 'help.category.preview3d',
-    icon: <HelpCategoryIcon paths={ICON_PATHS.cube} />,
-    shortcuts: [
-      {
-        keys: formatKey(SHORTCUTS.PREVIEW_TOGGLE).toUpperCase(),
-        descriptionKey: 'help.shortcut.togglePreview',
-      },
-      { keys: 'Space', descriptionKey: 'help.shortcut.expandPreview' },
-    ],
-  },
-  {
-    id: 'advanced',
-    nameKey: 'help.category.advanced',
-    icon: <HelpCategoryIcon paths={ICON_PATHS.settings} />,
-    shortcuts: [
-      {
-        keys: formatKey(SHORTCUTS.HALF_BIN_TOGGLE).toUpperCase(),
-        descriptionKey: 'help.shortcut.toggleHalfBin',
-      },
-      { keys: 'P', descriptionKey: 'help.shortcut.togglePaintMode' },
-    ],
-  },
-];
 
 export function HelpModal({ isOpen, onClose, isTablet = false }: HelpModalProps) {
   const t = useTranslation();
@@ -421,267 +265,5 @@ export function HelpModal({ isOpen, onClose, isTablet = false }: HelpModalProps)
         </div>
       </div>
     </div>
-  );
-}
-
-// Shortcut category section component
-function ShortcutCategorySection({
-  category,
-  modifierKey,
-}: {
-  category: ShortcutCategory;
-  modifierKey: string;
-}) {
-  const t = useTranslation();
-  return (
-    <section>
-      <div className="flex items-center gap-2 mb-3">
-        <span className="text-accent">{category.icon}</span>
-        <h3 style={STYLES.sectionHeader}>{t(category.nameKey)}</h3>
-      </div>
-      <div className="grid gap-2">
-        {category.shortcuts.map((shortcut, index) => (
-          <ShortcutRow
-            key={index}
-            keys={shortcut.keys}
-            description={t(shortcut.descriptionKey)}
-            modifier={shortcut.modifier}
-            shift={shortcut.shift}
-            modifierKey={modifierKey}
-          />
-        ))}
-      </div>
-    </section>
-  );
-}
-
-// Enhanced keyboard key component
-function KeyboardKey({ children }: { children: React.ReactNode }) {
-  return (
-    <kbd className="inline-flex items-center justify-center min-w-[24px] h-6 px-1.5 text-xs font-mono font-medium rounded border border-stroke bg-gradient-to-b from-surface-elevated to-surface text-content shadow-[0_1px_0_1px_rgba(0,0,0,0.2),inset_0_1px_0_rgba(255,255,255,0.1)]">
-      {children}
-    </kbd>
-  );
-}
-
-function ShortcutRow({
-  keys,
-  description,
-  modifier,
-  shift,
-  modifierKey,
-}: {
-  keys: string | readonly string[];
-  description: string;
-  modifier?: boolean;
-  shift?: boolean;
-  modifierKey: string;
-}) {
-  const keyArray = typeof keys === 'string' ? keys.split(' / ') : [...keys];
-
-  return (
-    <div className="flex justify-between items-center py-1.5 px-3 rounded-lg hover:bg-surface-hover/50 transition-colors">
-      <span className="text-sm text-content-secondary">{description}</span>
-      <div className="flex items-center gap-1">
-        {modifier && (
-          <>
-            <KeyboardKey>{modifierKey}</KeyboardKey>
-            <span className="text-content-tertiary text-xs">{KEY_SEPARATOR}</span>
-          </>
-        )}
-        {shift && (
-          <>
-            <KeyboardKey>Shift</KeyboardKey>
-            <span className="text-content-tertiary text-xs">{KEY_SEPARATOR}</span>
-          </>
-        )}
-        {keyArray.map((key, index) => (
-          <span key={index} className="flex items-center gap-1">
-            {index > 0 && <span className="text-content-tertiary text-xs mx-0.5">/</span>}
-            <KeyboardKey>{key}</KeyboardKey>
-          </span>
-        ))}
-      </div>
-    </div>
-  );
-}
-
-// Mouse interactions section
-function MouseInteractionsSection() {
-  const t = useTranslation();
-  return (
-    <section>
-      <div className="flex items-center gap-2 mb-3">
-        <span className="text-accent">
-          <HelpCategoryIcon paths={ICON_PATHS.mouse} />
-        </span>
-        <h3 style={STYLES.sectionHeader}>{t('help.mouse')}</h3>
-      </div>
-      <div className="grid gap-2">
-        <InteractionRow
-          action={t('help.mouse.clickDragEmpty')}
-          description={t('help.mouse.drawNewBin')}
-        />
-        <InteractionRow action={t('help.mouse.clickBin')} description={t('help.mouse.selectBin')} />
-        <InteractionRow
-          action={t('help.mouse.shiftClick')}
-          description={t('help.mouse.addToSelection')}
-        />
-        <InteractionRow
-          action={t('help.mouse.dragSelected')}
-          description={t('help.mouse.moveBins')}
-        />
-        <InteractionRow
-          action={t('help.mouse.altDragSelected')}
-          description={t('help.mouse.duplicateBins')}
-        />
-        <InteractionRow
-          action={t('help.mouse.dragEdges')}
-          description={t('help.mouse.resizeBin')}
-        />
-        <InteractionRow
-          action={t('help.mouse.doubleClickBin')}
-          description={t('help.mouse.quickLabelEdit')}
-        />
-        <InteractionRow
-          action={t('help.mouse.rightClickBin')}
-          description={t('help.mouse.contextMenu')}
-        />
-      </div>
-    </section>
-  );
-}
-
-// Touch gestures section
-function TouchGesturesSection() {
-  const t = useTranslation();
-  return (
-    <section>
-      <div className="flex items-center gap-2 mb-3">
-        <span className="text-accent">
-          <HelpCategoryIcon paths={ICON_PATHS.touch} />
-        </span>
-        <h3 style={STYLES.sectionHeader}>{t('help.touchGestures')}</h3>
-      </div>
-      <div className="grid gap-2">
-        <InteractionRow
-          action={t('help.gesture.tapBin')}
-          description={t('help.gesture.selectBin')}
-        />
-        <InteractionRow
-          action={t('help.gesture.dragEmpty')}
-          description={t('help.gesture.drawNewBin')}
-        />
-        <InteractionRow
-          action={t('help.gesture.dragSelected')}
-          description={t('help.gesture.moveBin')}
-        />
-        <InteractionRow
-          action={t('help.gesture.longPress')}
-          description={t('help.gesture.openContextMenu')}
-        />
-        <InteractionRow
-          action={t('help.gesture.dragEdge')}
-          description={t('help.gesture.resizeBin')}
-        />
-      </div>
-    </section>
-  );
-}
-
-function InteractionRow({ action, description }: { action: string; description: string }) {
-  return (
-    <div className="flex justify-between items-center py-1.5 px-3 rounded-lg hover:bg-surface-hover/50 transition-colors">
-      <span className="text-sm text-content-secondary">{description}</span>
-      <span className="text-xs text-content-tertiary bg-surface-elevated px-2 py-1 rounded">
-        {action}
-      </span>
-    </div>
-  );
-}
-
-// Tips section
-const TIP_KEYS = [
-  'help.tip.binPalette',
-  'help.tip.autoSplit',
-  'help.tip.dragLayers',
-  'help.tip.renameLayers',
-  'help.tip.autoSave',
-  'help.tip.quickOpen',
-  'help.tip.halfBin',
-] as const;
-
-function TipsSection() {
-  const t = useTranslation();
-
-  return (
-    <section>
-      <h3 className="mb-4" style={{ ...STYLES.sectionHeader, fontSize: 'var(--text-lg)' }}>
-        {t('help.tips')}
-      </h3>
-      <ul className="space-y-2 p-4 rounded-lg" style={STYLES.tipsList}>
-        {TIP_KEYS.map((key) => (
-          <li key={key} className="flex items-start gap-2">
-            <span style={STYLES.colorPrimary}>•</span>
-            <span>{t(key)}</span>
-          </li>
-        ))}
-      </ul>
-    </section>
-  );
-}
-
-// Blocked zones section
-function BlockedZonesSection() {
-  const t = useTranslation();
-  return (
-    <section>
-      <h3 className="mb-4" style={{ ...STYLES.sectionHeader, fontSize: 'var(--text-lg)' }}>
-        {t('help.blockedZones')}
-      </h3>
-      <div className="p-4 rounded-lg" style={STYLES.blockedZonesContent}>
-        <p className="mb-3">
-          <strong style={STYLES.textPrimary}>{t('help.whatAreBlockedZones')}</strong>
-        </p>
-        <p className="mb-3">{t('help.blockedZonesDescription')}</p>
-        <p>
-          <strong style={STYLES.textPrimary}>{t('help.example')}</strong>{' '}
-          {t('help.blockedZonesExample')}
-        </p>
-      </div>
-    </section>
-  );
-}
-
-// Bin clearance section
-function BinClearanceSection() {
-  const t = useTranslation();
-  return (
-    <section>
-      <h3 className="mb-4" style={{ ...STYLES.sectionHeader, fontSize: 'var(--text-lg)' }}>
-        {t('help.binClearance')}
-      </h3>
-      <div className="p-4 rounded-lg" style={STYLES.blockedZonesContent}>
-        <p className="mb-3">
-          <strong style={STYLES.textPrimary}>{t('help.whatIsClearance')}</strong>
-        </p>
-        <p className="mb-3">{t('help.clearanceDescription')}</p>
-        <p className="mb-3">
-          <strong style={STYLES.textPrimary}>{t('help.example')}</strong>{' '}
-          {t('help.clearanceExample')}
-        </p>
-        <p>{t('help.clearanceHowTo')}</p>
-      </div>
-    </section>
-  );
-}
-
-function HelpCategoryIcon({ paths }: { paths: readonly string[] }) {
-  return (
-    <svg className="w-4 h-4" fill="none" viewBox="0 0 24 24" stroke="currentColor">
-      {paths.map((d) => (
-        <path key={d} strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d={d} />
-      ))}
-    </svg>
   );
 }
